Prevent drag & drop between conditions & actions
